UI creation and JS addition
a) In the UI, the player is able to play the game by clicking on buttons rather than
typing their answer in a prompt.
b) Created three buttons, one for each selection. Added an event listener to the buttons that call the playRound function with the correct playerSelection every time a button is clicked.
c) Added a div for displaying results and changed all of the console.logs into DOM methods.
d) Displayed the running score, and announced a winner of the game once one player reaches 5 points.